Replace old sync module with the new one
